+++ test/bigtest/network/models/comments.js
@@ -1,0 +1,4 @@
+import { Model } from '@bigtest/mirage';
+
+export default Model.extend({
+});